GenAI_Projects/Proj-4_TelegramBot/requirements.txt
aiogram
openai==0.28 
python-dotenv